refactor: ticket 패키지 분리
