+++ src/main/java/lotto/controller/LottoController.java
@@ -7,8 +7,8 @@ import lotto.domain.number.SerialLottoNumber;
 import lotto.domain.number.SerialLottoNumberFactory;
 import lotto.domain.random.RandomLottoNumberGenerator;
 import lotto.domain.result.LottoResult;
-import lotto.domain.result.LottoTickets;
-import lotto.domain.result.LottoTicketsFactory;
+import lotto.domain.ticket.LottoTickets;
+import lotto.domain.ticket.LottoTicketsFactory;
 import lotto.domain.result.Winning;
 import lotto.exceptions.*;
 import lotto.view.InputView;
